The one that removed eucarc.
Instead of passing a eucarc file the setup now takes a CLC ip address. The test runner generates a set of cloud admin credentials (if creds created by the runner are not already present) and will retrieve that ini file. From this file endpoints and credential info is parsed.

There is a new basic test of functionality (N4jTest) that demonstrates getting cloud connection and cleaning up credentials after the test run.

Some testing was done to ensure existing tests are in working order but there are probably some that will still need to be updated for the changes.