Update S3ObjectACLTests.java
The fix for EUCA-12628 removes the default fullcontrol grant for the obect owner which causes putObject_ACL_Headers() to fail. The problem is that without the old fullcontol grant getAcl fails (no permission) and the number of grants is wrong (Owner fullcontrol is no longer there)in the acl verification. The solution was to manually add the owner grant before the calls to putObjectWithACL